requirements.txt
# === Core Dependencies for AI Log Intelligence System ===

# RabbitMQ client library — used by Producer and Consumer to publish/consume messages
pika==1.3.2

# High-performance async web framework — powers the REST API layer
fastapi==0.115.6

# ASGI server — runs the FastAPI application in production
uvicorn[standard]==0.34.0

# OpenSearch Python client — indexes and queries logs in OpenSearch
opensearch-py==2.8.0

# Loads environment variables from .env files — keeps secrets out of code
python-dotenv==1.0.1

# JSON schema validation — validates log payloads before processing
pydantic==2.10.4

# HTTP client library — used for health checks and potential LLM API calls
requests==2.32.3
